add kotlin part 1
