--- app/src/main/java/com/almaz/task1/classes_block/ClassesBlock.java
@@ -1,76 +1,0 @@
-package com.almaz.task1.classes_block;
-
-/**
- * Набор заданий по работе с классами в java.
- * <p>
- * Задания подразумевают создание класса(ов), выполняющих задачу.
- * <p>
- * Проверка осуществляется ментором.
- */
-public interface ClassesBlock {
-
-    /*
-      I
-
-      Создать класс с двумя переменными. Добавить функцию вывода на экран
-      и функцию изменения этих переменных. Добавить функцию, которая находит
-      сумму значений этих переменных, и функцию, которая находит наибольшее
-      значение из этих двух переменных.
-     */
-
-    /*
-      II
-
-      Создать класс, содержащий динамический массив и количество элементов в нем.
-      Добавить конструктор, который выделяет память под заданное количество элементов.
-      Добавить методы, позволяющие заполнять массив случайными числами,
-      переставлять в данном массиве элементы в случайном порядке, находить количество
-      различных элементов в массиве, выводить массив на экран.
-     */
-
-    /*
-      III
-
-      Описать класс, представляющий треугольник. Предусмотреть методы для создания объектов,
-      вычисления площади, периметра и точки пересечения медиан.
-      Описать свойства для получения состояния объекта.
-     */
-
-    /*
-      IV
-
-      Составить описание класса для представления времени.
-      Предусмотреть возможности установки времени и изменения его отдельных полей
-      (час, минута, секунда) с проверкой допустимости вводимых значений.
-      В случае недопустимых значений полей выбрасываются исключения.
-      Создать методы изменения времени на заданное количество часов, минут и секунд.
-     */
-
-    /*
-      V
-
-      Класс Абонент: Идентификационный номер, Фамилия, Имя, Отчество, Адрес,
-      Номер кредитной карточки, Дебет, Кредит, Время междугородных и городских переговоров;
-      Конструктор; Методы: установка значений атрибутов, получение значений атрибутов,
-      вывод информации. Создать массив объектов данного класса.
-      Вывести сведения относительно абонентов, у которых время городских переговоров
-      превышает заданное.  Сведения относительно абонентов, которые пользовались
-      междугородной связью. Список абонентов в алфавитном порядке.
-     */
-
-    /*
-      VI
-
-      Задача на взаимодействие между классами. Разработать систему «Вступительные экзамены».
-      Абитуриент регистрируется на Факультет, сдает Экзамены. Преподаватель выставляет Оценку.
-      Система подсчитывает средний бал и определяет Абитуриента, зачисленного в учебное заведение.
-     */
-
-    /*
-      VII
-
-      Задача на взаимодействие между классами. Разработать систему «Интернет-магазин».
-      Товаровед добавляет информацию о Товаре. Клиент делает и оплачивает Заказ на Товары.
-      Товаровед регистрирует Продажу и может занести неплательщика в «черный список».
-     */
-}
