--- app/src/main/java/com/almaz/task1/classes_block/entry_exams/Person.java
@@ -1,37 +1,0 @@
-package com.almaz.task1.classes_block.entry_exams;
-
-import androidx.annotation.NonNull;
-
-public abstract class Person {
-    private String firstName;
-    private String lastName;
-
-    public Person(String firstName, String lastName) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    @NonNull
-    @Override
-    public String toString() {
-        return lastName + " " + firstName;
-    }
-
-
-}
