+++ app/src/main/java/com/almaz/task1/java_training/part_1/ArraysTraining.java
@@ -1,0 +1,120 @@
+package com.almaz.task1.java_training.part_1;
+
+public class ArraysTraining {
+
+    /**
+     * Метод должен сортировать входящий массив
+     * по возрастранию пузырьковым методом
+     *
+     * @param valuesArray массив для сортировки
+     */
+    public int[] sort(int[] valuesArray) {
+        int temporaryValue;
+
+        for (int i = 0; i < valuesArray.length; i++) {
+            for (int j = 0; j < valuesArray.length - 1; j++) {
+                if (valuesArray[j] > valuesArray[j + 1]) {
+                    temporaryValue = valuesArray[j + 1];
+                    valuesArray[j + 1] = valuesArray[j];
+                    valuesArray[j] = temporaryValue;
+                }
+            }
+        }
+
+        return valuesArray;
+    }
+
+    /**
+     * Метод должен возвращать максимальное
+     * значение из введенных. Если входящие числа
+     * отсутствуют - вернуть 0
+     *
+     * @param values входящие числа
+     * @return максимальное число или 0
+     */
+    public int maxValue(int... values) {
+        int maxValue;
+
+        if (values.length == 0) {
+            maxValue = 0;
+        } else {
+            maxValue = values[0];
+            for (int value : values) {
+                maxValue = Math.max(value, maxValue);
+            }
+        }
+
+        return maxValue;
+    }
+
+    /**
+     * Переставить элементы массива
+     * в обратном порядке
+     *
+     * @param array массив для преобразования
+     * @return входящий массив в обратном порядке
+     */
+    public int[] reverse(int[] array) {
+        int temporaryValue;
+
+        for (int i = 0; i < array.length / 2; i++) {
+            temporaryValue = array[i];
+            array[i] = array[array.length - i - 1];
+            array[array.length - i - 1] = temporaryValue;
+        }
+
+        return array;
+    }
+
+    /**
+     * Метод должен вернуть массив,
+     * состоящий из чисел Фибоначчи
+     *
+     * @param numbersCount количество чисел Фибоначчи,
+     *                     требуемое в исходящем массиве.
+     *                     Если numbersCount < 1, исходный
+     *                     массив должен быть пуст.
+     * @return массив из чисел Фибоначчи
+     */
+    public int[] fibonacciNumbers(int numbersCount) {
+        int[] fibonacciArray;
+
+        if (numbersCount < 1) {
+            fibonacciArray = new int[]{};
+        } else {
+            fibonacciArray = new int[numbersCount];
+            fibonacciArray[0] = 1;
+            fibonacciArray[1] = 1;
+            for (int i = 2; i < numbersCount; i++) {
+                fibonacciArray[i] = fibonacciArray[i - 1] + fibonacciArray[i - 2];
+            }
+        }
+
+        return fibonacciArray;
+    }
+
+    /**
+     * В данном массиве найти максимальное
+     * количество одинаковых элементов.
+     *
+     * @param array массив для выборки
+     * @return количество максимально встречающихся
+     * элементов
+     */
+    public int maxCountSymbol(int[] array) {
+        int maxCount = 0;
+        int counter;
+
+        for (int comparedNumber  : array) {
+            counter = 0;
+            for (int numberFromComparableArray : array) {
+                if (comparedNumber == numberFromComparableArray) {
+                    counter++;
+                }
+                if (counter > maxCount) maxCount = counter;
+            }
+        }
+
+        return maxCount;
+    }
+}
