+++ app/src/main/java/com/almaz/task1/java_training/part_1/ElementaryTraining.java
@@ -1,0 +1,107 @@
+package com.almaz.task1.java_training.part_1;
+
+/**
+ * Набор тренингов по работе с примитивными типами java.
+ * <p>
+ * Задания определены в комментариях методов.
+ * <p>
+ * Проверка может быть осуществлена запуском тестов.
+ * <p>
+ * Доступна проверка тестированием @see ElementaryTrainingTest.
+ */
+public class ElementaryTraining {
+
+    /**
+     * Метод должен возвращать среднее значение
+     * для введенных параметров
+     *
+     * @param firstValue  первый элемент
+     * @param secondValue второй элемент
+     * @return среднее значение для введенных чисел
+     */
+    public double averageValue(int firstValue, int secondValue) {
+        return (firstValue + secondValue) / 2.0;
+    }
+
+    /**
+     * Пользователь вводит три числа.
+     * Произвести манипуляции и вернуть сумму новых чисел
+     *
+     * @param firstValue  увеличить в два раза
+     * @param secondValue уменьшить на три
+     * @param thirdValue  возвести в квадрат
+     * @return сумма новых трех чисел
+     */
+    public double complicatedAmount(int firstValue, int secondValue, int thirdValue) {
+        return firstValue * 2 + (secondValue - 3) + Math.pow(thirdValue, 2);
+    }
+
+    /**
+     * Метод должен поменять значение в соответствии с условием.
+     * Если значение больше 3, то увеличить
+     * на 10, иначе уменьшить на 10.
+     *
+     * @param value число для изменения
+     * @return новое значение
+     */
+    public int changeValue(int value) {
+        return (value > 3) ? value + 10 : value - 10;
+    }
+
+    /**
+     * Метод должен менять местами первую
+     * и последнюю цифру числа.
+     * Обрабатывать максимум пятизначное число.
+     * Если число < 10, вернуть
+     * то же число
+     *
+     * @param value число для перестановки
+     * @return новое число
+     */
+    public int swapNumbers(int value) {
+        String textValue = String.valueOf(value);
+        int result;
+
+        if (value < 10 || textValue.length() > 5) {
+            result = value;
+        } else {
+            textValue = textValue.charAt(textValue.length() - 1)
+                    + textValue.substring(1, textValue.length() - 1)
+                    + textValue.charAt(0);
+            result = Integer.parseInt(textValue);
+        }
+
+        return result;
+    }
+
+    /**
+     * Изменить значение четных цифр числа на ноль.
+     * Счет начинать с единицы.
+     * Обрабатывать максимум пятизначное число.
+     * Если число < 10 вернуть
+     * то же число.
+     *
+     * @param value число для изменения
+     * @return новое число
+     */
+    public int zeroEvenNumber(int value) {
+        String textValue = String.valueOf(value);
+        StringBuilder stringBuilder = new StringBuilder(textValue);
+        int result;
+
+        if (value < 10 || textValue.length() > 5) {
+            result = value;
+        } else {
+            for (int i = 0; i < textValue.length(); i++) {
+                if (textValue.charAt(i) % 2 == 0) {
+                    stringBuilder.replace(i, i+1, "0");
+                    textValue = stringBuilder.toString();
+                }
+            }
+            result = Integer.parseInt(textValue);
+        }
+
+        return result;
+    }
+}
+
