+++ app/src/main/java/com/almaz/task1/java_training/part_2/classes_block/internet_store/Person.java
@@ -1,0 +1,37 @@
+package com.almaz.task1.java_training.part_2.classes_block.internet_store;
+
+import androidx.annotation.NonNull;
+
+public abstract class Person {
+    private String firstName;
+    private String lastName;
+
+    public Person(String firstName, String lastName) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    @NonNull
+    @Override
+    public String toString() {
+        return lastName + " " + firstName;
+    }
+
+
+}
