+++ app/src/main/java/com/almaz/task1/java_training/part_2/classes_block/internet_store/Product.java
@@ -1,0 +1,49 @@
+package com.almaz.task1.java_training.part_2.classes_block.internet_store;
+
+import androidx.annotation.NonNull;
+import java.util.Objects;
+
+public class Product {
+    private String name;
+    private double price;
+
+    public Product(String name, double price) {
+        this.name = name;
+        this.price = price;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public double getPrice() {
+        return price;
+    }
+
+    public void setPrice(double price) {
+        this.price = price;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof Product)) return false;
+        Product product = (Product) o;
+        return Double.compare(product.price, price) == 0 && name.equals(product.name);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name, price);
+    }
+
+    @NonNull
+    @Override
+    public String toString() {
+        return name;
+    }
+}
