+++ app/src/main/java/com/almaz/task1/java_training/part_2/classes_block/triangle/Point.java
@@ -1,0 +1,27 @@
+package com.almaz.task1.java_training.part_2.classes_block.triangle;
+
+public class Point {
+    private double x;
+    private double y;
+
+    public Point(double x, double y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public double getX() {
+        return x;
+    }
+
+    public void setX(double x) {
+        this.x = x;
+    }
+
+    public double getY() {
+        return y;
+    }
+
+    public void setY(double y) {
+        this.y = y;
+    }
+}
