+++ app/src/main/java/com/almaz/task1/kotlin_training/part_1/Magazine.kt
@@ -1,0 +1,11 @@
+package com.almaz.task1.kotlin_training.part_1
+
+class Magazine(
+    override val price: Long,
+    override val wordCount: Long
+) : Publication {
+
+    override fun getType(): String {
+        return MAGAZINE
+    }
+}
