+++ app/src/test/java/com/almaz/task1/ArraysTrainingTest.java
@@ -3,6 +3,8 @@ package com.almaz.task1;
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 
+import com.almaz.task1.java_training.part_1.ArraysTraining;
+
 import org.junit.BeforeClass;
 import org.junit.Test;
